+++ mne_connectivity/base.py
@@ -1,23 +1,73 @@
+from copy import copy
+
 import numpy as np
 import xarray as xr
+from mne.utils import (_check_combine, _check_option, _validate_type,
+                       copy_function_doc_to_method_doc, object_size,
+                       sizeof_fmt)
 
-from mne.utils import (sizeof_fmt, object_size,
-                       _validate_type, _check_option,
-                       copy_function_doc_to_method_doc)
 from mne_connectivity.utils import fill_doc
-from mne_connectivity.viz import (plot_connectivity_circle)
+from mne_connectivity.viz import plot_connectivity_circle
 
 
 class SpectralMixin:
     @property
     def freqs(self):
+        """The frequency points of the connectivity data.
+
+        If these are computed over a frequency band, it will
+        be the median frequency of the frequency band.
+        """
         return self.xarray.coords.get('freqs').values.tolist()
 
 
 class TimeMixin:
     @property
     def times(self):
+        """The time points of the connectivity data."""
         return self.xarray.coords.get('times').values.tolist()
+
+
+class EpochMixin:
+    def combine(self, combine='mean'):
+        """Combine connectivity data over epochs.
+
+        Parameters
+        ----------
+        combine : 'mean' | 'median' | callable
+            How to combine correlation estimates across epochs.
+            Default is 'mean'. If callable, it must accept one
+            positional input. For example::
+
+                combine = lambda data: np.median(data, axis=0)
+
+        Returns
+        -------
+        conn : instance of Connectivity
+            The combined connectivity data structure.
+        """
+        from .io import _xarray_to_conn
+
+        fun = _check_combine(combine, valid=('mean', 'median'))
+
+        # apply function over the dataset
+        new_xr = xr.apply_ufunc(fun, self.xarray,
+                                input_core_dims=[['epochs']],
+                                vectorize=True)
+        new_xr.attrs = self.xarray.attrs
+
+        # map class name to its actual class
+        conn_cls = {
+            'EpochConnectivity': Connectivity,
+            'EpochTemporalConnectivity': TemporalConnectivity,
+            'EpochSpectralConnectivity': SpectralConnectivity,
+            'EpochSpectroTemporalConnectivity': SpectroTemporalConnectivity
+        }
+        cls_func = conn_cls[self.__class__.__name__]
+
+        # convert new xarray to non-Epoch data structure
+        conn = _xarray_to_conn(new_xr, cls_func)
+        return conn
 
 
 @fill_doc
@@ -54,6 +104,12 @@ class _Connectivity():
     full connectivity structure is computed, or a subset of the nodes
     (equal to the length of the indices passed in).
 
+    Since we store connectivity data as a raveled array, one can
+    easily optimize the storage of "symmetric" connectivity data.
+    One can use numpy to convert a full all-to-all connectivity
+    into an upper triangular portion, and set ``indices='symmetric'``.
+    This would reduce the RAM needed in half.
+
     The underlying data structure is an ``xarray.DataArray``,
     with a similar API to ``xarray``. We provide support for storing
     connectivity data in a subset of nodes. Thus the underlying
@@ -68,18 +124,16 @@ class _Connectivity():
                  n_nodes, **kwargs):
 
         if isinstance(indices, str) and \
-                indices not in ['all']:
+                indices not in ['all', 'symmetric']:
             raise ValueError(f'Indices can only be '
                              f'"all", otherwise '
                              f'should be a list of tuples. '
                              f'It cannot be {indices}.')
 
         # check the incoming data structure
-        self.method = method
-        self.indices = indices
-        self.n_nodes = n_nodes
-        self._check_data_consistency(data)
-        self._prepare_xarray(data, names=names, **kwargs)
+        self._check_data_consistency(data, indices=indices, n_nodes=n_nodes)
+        self._prepare_xarray(data, names=names, indices=indices,
+                             n_nodes=n_nodes, method=method, **kwargs)
 
     def __repr__(self) -> str:
         r = f'<{self.__class__.__name__} | '
@@ -88,9 +142,9 @@ class _Connectivity():
             r += "freq : [%f, %f], " % (self.freqs[0], self.freqs[-1])
         if 'times' in self.dims:
             r += "time : [%f, %f], " % (self.times[0], self.times[-1])
-        r += ", nave : %d" % self.n_epochs_used
-        r += ', nodes, n_estimated : %d, %d' % (self.n_nodes,
-                                                self.n_estimated_nodes)
+        r += f", nave : {self.n_epochs_used}"
+        r += f', nodes, n_estimated : {self.n_nodes}, ' \
+             f'{self.n_estimated_nodes}'
         r += ', ~%s' % (sizeof_fmt(self._size),)
         r += '>'
         return r
@@ -106,9 +160,8 @@ class _Connectivity():
             start_idx = 0
         self.n_estimated_nodes = data.shape[start_idx]
 
-        return data
-
-    def _prepare_xarray(self, data, names, **kwargs):
+    def _prepare_xarray(self, data, names, indices, n_nodes, method,
+                        **kwargs):
         """Prepare xarray data structure.
 
         Parameters
@@ -118,12 +171,9 @@ class _Connectivity():
         names : [type]
             [description]
         """
-        # get the number of estimated nodes
-        data = self._get_n_estimated_nodes(data)
-
         # set node names
         if names is None:
-            names = list(map(str, range(self.n_nodes)))
+            names = list(map(str, range(n_nodes)))
 
         # the names of each first few dimensions of
         # the data depending if data is epoched or not
@@ -154,6 +204,14 @@ class _Connectivity():
                 kwargs[key] = val.tolist()
         kwargs['node_names'] = names
 
+        # set method, indices and n_nodes
+        if isinstance(indices, tuple):
+            new_indices = (list(indices[0]), list(indices[1]))
+            indices = new_indices
+        kwargs['method'] = method
+        kwargs['indices'] = indices
+        kwargs['n_nodes'] = n_nodes
+
         # create xarray object
         xarray_obj = xr.DataArray(
             data=data,
@@ -163,7 +221,7 @@ class _Connectivity():
         )
         self._obj = xarray_obj
 
-    def _check_data_consistency(self, data):
+    def _check_data_consistency(self, data, indices, n_nodes):
         """Perform data input checks."""
         if not isinstance(data, np.ndarray):
             raise TypeError('Connectivity data must be passed in as a '
@@ -184,14 +242,30 @@ class _Connectivity():
                                    f'dimensions. Your data was '
                                    f'{data.shape} shape.')
 
+        # get the number of estimated nodes
+        self._get_n_estimated_nodes(data)
+        if self.is_epoched:
+            data_len = data.shape[1]
+        else:
+            data_len = data.shape[0]
+
         # check that the indices passed in are of the same length
-        if isinstance(self.indices, tuple):
-            if len(self.indices[0]) != len(self.indices[1]):
+        if isinstance(indices, tuple):
+            if len(indices[0]) != len(indices[1]):
                 raise ValueError(f'If indices are passed in '
                                  f'then they must be the same '
                                  f'length. They are right now '
-                                 f'{len(self.indices[0])} and '
-                                 f'{len(self.indices[1])}.')
+                                 f'{len(indices[0])} and '
+                                 f'{len(indices[1])}.')
+        elif indices == 'symmetric':
+            expected_len = ((n_nodes + 1) * n_nodes) // 2
+            if data_len != expected_len:
+                raise ValueError(f'If "indices" is "symmetric", then '
+                                 f'connectivity data should be the '
+                                 f'upper-triangular part of the matrix. There '
+                                 f'are {data_len} estimated connections. '
+                                 f'But there should be {expected_len} '
+                                 f'estimated connections.')
 
     @property
     def _data(self):
@@ -220,6 +294,35 @@ class _Connectivity():
     def shape(self):
         """Shape of raveled connectivity."""
         return self.xarray.shape
+
+    @property
+    def n_nodes(self):
+        """The number of nodes in the dataset.
+
+        Even if ``indices`` defines a subset of nodes that
+        were computed, this should be the total number of
+        nodes in the original dataset.
+        """
+        return self.attrs['n_nodes']
+
+    @property
+    def method(self):
+        """The method used to compute connectivity."""
+        return self.attrs['method']
+
+    @property
+    def indices(self):
+        """Indices of connectivity data.
+
+        Returns
+        -------
+        indices : str | tuple of lists
+            Either 'all' for all-to-all connectivity,
+            'symmetric' for symmetric all-to-all connectivity,
+            or a tuple of lists representing the node-to-nodes
+            that connectivity was computed.
+        """
+        return self.attrs['indices']
 
     @property
     def names(self):
@@ -293,6 +396,18 @@ class _Connectivity():
                     data[:, row_idx, col_idx, ...] = self._data
                 else:
                     data[row_idx, col_idx, ...] = self._data
+            elif self.indices == 'symmetric':
+                data = np.zeros(new_shape)
+
+                # get the upper/lower triangular indices
+                row_triu_inds, col_triu_inds = np.triu_indices(
+                    self.n_nodes, k=0)
+                if self.is_epoched:
+                    data[:, row_triu_inds, col_triu_inds, ...] = self._data
+                    data[:, col_triu_inds, row_triu_inds, ...] = self._data
+                else:
+                    data[row_triu_inds, col_triu_inds, ...] = self._data
+                    data[col_triu_inds, row_triu_inds, ...] = self._data
             else:
                 data = self._data.reshape(new_shape)
 
@@ -354,8 +469,43 @@ class _Connectivity():
     # def plot_3d(self):
     #     pass
 
-    # def save(self, fname):
-    #     pass
+    def save(self, fname):
+        """Save connectivity data to disk.
+
+        Parameters
+        ----------
+        fname : str | pathlib.Path
+            The filepath to save the data. Data is saved
+            as netCDF files (``.nc`` extension).
+        """
+        method = self.method
+        indices = self.indices
+        n_nodes = self.n_nodes
+
+        # create a copy of the old attributes
+        old_attrs = copy(self.attrs)
+
+        # assign these to xarray's attrs
+        self.attrs['method'] = method
+        self.attrs['indices'] = indices
+        self.attrs['n_nodes'] = n_nodes
+
+        # save the name of the connectivity structure
+        self.attrs['data_structure'] = str(self.__class__.__name__)
+
+        # netCDF does not support 'None'
+        # so map these to 'n/a'
+        for key, val in self.attrs.items():
+            if val is None:
+                self.attrs[key] = 'n/a'
+
+        # save as a netCDF file
+        # note this requires the netcdf4 python library
+        self.xarray.to_netcdf(fname, mode='w', format='NETCDF4',
+                              engine='netcdf4')
+
+        # re-set old attributes
+        self.xarray.attrs = old_attrs
 
 
 @fill_doc
@@ -375,7 +525,7 @@ class SpectralConnectivity(_Connectivity, SpectralMixin):
     """
 
     def __init__(self, data, freqs, n_nodes, names=None,
-                 indices=None, method=None, spec_method=None,
+                 indices='all', method=None, spec_method=None,
                  n_epochs_used=None, **kwargs):
         super().__init__(data, names=names, method=method,
                          indices=indices, n_nodes=n_nodes,
@@ -398,7 +548,7 @@ class TemporalConnectivity(_Connectivity, TimeMixin):
     %(n_epochs_used)s
     """
 
-    def __init__(self, data, times, n_nodes, names=None, indices=None,
+    def __init__(self, data, times, n_nodes, names=None, indices='all',
                  method=None, n_epochs_used=None, **kwargs):
         super().__init__(data, names=names, method=method,
                          n_nodes=n_nodes, indices=indices,
@@ -424,7 +574,7 @@ class SpectroTemporalConnectivity(_Connectivity, SpectralMixin, TimeMixin):
     """
 
     def __init__(self, data, freqs, times, n_nodes, names=None,
-                 indices=None, method=None,
+                 indices='all', method=None,
                  spec_method=None, n_epochs_used=None, **kwargs):
         super().__init__(data, names=names, method=method, indices=indices,
                          n_nodes=n_nodes, freqs=freqs,
@@ -433,7 +583,7 @@ class SpectroTemporalConnectivity(_Connectivity, SpectralMixin, TimeMixin):
 
 
 @fill_doc
-class EpochSpectralConnectivity(SpectralConnectivity):
+class EpochSpectralConnectivity(SpectralConnectivity, EpochMixin):
     """Spectral connectivity container over Epochs.
 
     Parameters
@@ -450,7 +600,7 @@ class EpochSpectralConnectivity(SpectralConnectivity):
     is_epoched = True
 
     def __init__(self, data, freqs, n_nodes, names=None,
-                 indices=None, method=None,
+                 indices='all', method=None,
                  spec_method=None, **kwargs):
         super().__init__(
             data, freqs=freqs, names=names, indices=indices,
@@ -459,7 +609,7 @@ class EpochSpectralConnectivity(SpectralConnectivity):
 
 
 @fill_doc
-class EpochTemporalConnectivity(TemporalConnectivity):
+class EpochTemporalConnectivity(TemporalConnectivity, EpochMixin):
     """Temporal connectivity container over Epochs.
 
     Parameters
@@ -475,14 +625,16 @@ class EpochTemporalConnectivity(TemporalConnectivity):
     is_epoched = True
 
     def __init__(self, data, times, n_nodes, names=None,
-                 indices=None, method=None, **kwargs):
+                 indices='all', method=None, **kwargs):
         super().__init__(data, times=times, names=names,
                          indices=indices, n_nodes=n_nodes,
                          method=method, **kwargs)
 
 
 @fill_doc
-class EpochSpectroTemporalConnectivity(SpectroTemporalConnectivity):
+class EpochSpectroTemporalConnectivity(
+    SpectroTemporalConnectivity, EpochMixin
+):
     """Spectrotemporal connectivity container over Epochs.
 
     Parameters
@@ -500,7 +652,7 @@ class EpochSpectroTemporalConnectivity(SpectroTemporalConnectivity):
     is_epoched = True
 
     def __init__(self, data, freqs, times, n_nodes,
-                 names=None, indices=None, method=None,
+                 names=None, indices='all', method=None,
                  spec_method=None, **kwargs):
         super().__init__(
             data, names=names, freqs=freqs, times=times, indices=indices,
@@ -509,7 +661,7 @@ class EpochSpectroTemporalConnectivity(SpectroTemporalConnectivity):
 
 
 @fill_doc
-class Connectivity(_Connectivity):
+class Connectivity(_Connectivity, EpochMixin):
     """Connectivity container without frequency or time component.
 
     Parameters
@@ -522,7 +674,7 @@ class Connectivity(_Connectivity):
     %(n_epochs_used)s
     """
 
-    def __init__(self, data, n_nodes, names=None, indices=None,
+    def __init__(self, data, n_nodes, names=None, indices='all',
                  method=None, n_epochs_used=None, **kwargs):
         super().__init__(data, names=names, method=method,
                          n_nodes=n_nodes, indices=indices,
@@ -531,7 +683,7 @@ class Connectivity(_Connectivity):
 
 
 @fill_doc
-class EpochConnectivity(_Connectivity):
+class EpochConnectivity(_Connectivity, EpochMixin):
     """Epoch connectivity container.
 
     Parameters
@@ -547,7 +699,7 @@ class EpochConnectivity(_Connectivity):
     # whether or not the connectivity occurs over epochs
     is_epoched = True
 
-    def __init__(self, data, n_nodes, names=None, indices=None,
+    def __init__(self, data, n_nodes, names=None, indices='all',
                  method=None, n_epochs_used=None, **kwargs):
         super().__init__(data, names=names, method=method,
                          n_nodes=n_nodes, indices=indices,