+++ mne_connectivity/tests/test_utils.py
@@ -1,8 +1,8 @@
 import numpy as np
-from numpy.testing import assert_array_equal
 import pytest
+from numpy.testing import assert_array_equal
 
-from mne_connectivity.utils import seed_target_indices, degree
+from mne_connectivity.utils import degree, seed_target_indices
 
 
 def test_indices():