+++ mne_connectivity/viz/__init__.py
@@ -1,4 +1,4 @@
 """Visualization routines."""
 
-from .circle import plot_connectivity_circle, circular_layout
 from ._3d import plot_sensors_connectivity
+from .circle import circular_layout, plot_connectivity_circle